xinitrc.auto is a bit more flexible now
Now has a configurable list of 'preferred' and 'fallback' sessions, and
looks for other sessions in /etc/X11/xinit. Preferred sessions are tried
first, then sessions in /etc/X11/xinit (excluding sessions in the
fallback list) in order of most recent modification dates, and finally
the fallback sessions in order.