X welcome dialog now has shutdown/restart buttons
This is admittedly janky. Internally, the function that starts X as the
slackware user creates a pipe that only slakresq can access, and listens
for the reboot/shutdown command from it.

The pipe is created go-rwx, so it should never be in a state where
another hypothetical user can write to it.